READ extended description
made install, search, and change-repo flags their own scripts to be sourced by pacstall. It is to overwhelming to work on a single script so splitting it up makes it easier to work and not mess up other important parts. The rest of the flags will remain in the pacstall script as their very trivial.

Former-commit-id: 7df780785735c8e4f2e15ab2a20dc9395fca3d44
Former-commit-id: 621a56a02c4f058cf98a7fae117078d87b56d549