Removed pdb because porg works fine and existing tools were better than what I had. Two important changes to pacscripts
- destdir
- bindir

destdir is where packages are installed. The standard is
/usr/src/pacstall unless something breaks. bindir is like a mini PATH.
For instance, if I installed kvantum to /usr/src/pacstall, it will
create the kvantum directory and put the contents into
/usr/src/pacstall/kvantum/. If kvantum were installed globally, it's
binaries would be put in /usr/local/bin, but since we are isolating it,
they are in /usr/src/pacstall/kvantum/usr/local/bin so our bindir is
usr/local/bin. Pacstall will simlink to /usr/local/bin so files are
isolated and easier to track while still being in the users PATH


Former-commit-id: 8dfa032a784f98dcd45e30553a5de0ef6ab3ef35
Former-commit-id: e2d686a8a3260300f21a23211fd6bbb36d5da554