Added hashing feature
To generate a hash for your package, download the file as it is, don't
rename or anything, and run sha256sum <file> | sed 's/\s.*$//'. That
will take out the file, we just want the string, or alternatively, just
run sha256sum <file> and copy the string. Put that into your pacscript
under the variable hash, and your ready to go! Pacstall will generate a
hash for the package and check that against the hash variable.


Former-commit-id: c0980434679f2a6db40aa53346b79c705fbe4ba0 [formerly 36697b6d8b7fc3dc49e83851ac6114acf3b3007b] [formerly 2282df9318efd382f292eb80c468142b96d8b532]
Former-commit-id: d44f7e6ad2d69c0ad33eac98ffddd0138416989d
Former-commit-id: 54aec35e6c8464ad92c63ab45ff3e0e659faf94b