Added hashing feature
To generate a hash for your package, download the file as it is, don't
rename or anything, and run sha256sum <file> | sed 's/\s.*$//'. That
will take out the file, we just want the string, or alternatively, just
run sha256sum <file> and copy the string. Put that into your pacscript
under the variable hash, and your ready to go! Pacstall will generate a
hash for the package and check that against the hash variable.


Former-commit-id: 2282df9318efd382f292eb80c468142b96d8b532