Feat: srcdir and pkgdir added
`srcdir` is the package's source directory. For example, pkg foo has a
directory that is extracted to `/tmp/pacstall` called `foo-1.0`. The
`srcdir` would be `/tmp/pacstall/foo-1.0`

`pkgdir` is the path that a package is to be installed to. Package
`foo`'s `pkgdir` would be `/usr/src/pacstall/foo`