tempfix: force set `$URL` to repair `-Up` function
Original issue: `rhino-core` received a dependency update for `unicorn-desktop`, but `unicorn-desktop` was added to the repository after the previous `rhino-core` update. As a result, `-Up` fails to update `rhino-core` from beta4 to beta5, and an `-I` to fetch the repo must be done to update.

After looking into it, we found that `$URL` is getting eaten up by something, and not outputting correctly. We have found this fix to work, which subsidizes the repo into the `$URL` directly.