Ensure files and directories created with sudo are world-readable
I run my system with a 0007 umask, which means some of the files and directories created by pacstall with sudo aren't then readable by it when it returns back to running as a normal user.

This patch fixes the two locations I've found so far, but I suspect there are others.